+++ build.gradle
@@ -1,6 +1,6 @@
 allprojects {
     group 'com.incarcloud'
-    version '0.6.0-SNAPSHOT'
+    version '0.6.1-SNAPSHOT'
 }
 
 subprojects {
