+++ sb2-web-plat/build.gradle
@@ -3,4 +3,6 @@ apply plugin: 'org.springframework.boot'
 dependencies {
     compile project(":sb2-web")
     compile project(":sb2-service")
+
+    compile("org.springframework.boot:spring-boot-starter-security:${springBootVersion}")
 }
