+++ sb2-web-plat/src/main/java/com/incarcloud/sb2/auth/api/AuthApi.java
@@ -1,6 +1,6 @@
 package com.incarcloud.sb2.auth.api;
 
-import com.incarcloud.sb2.security.LoginUserInfo;
+import com.incarcloud.mvc.security.entity.LoginUserInfo;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
@@ -8,6 +8,7 @@ import io.swagger.annotations.ApiOperation;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.Map;
 
 /**
@@ -26,14 +27,14 @@ public interface AuthApi {
     @PostMapping("/login")
     default void fakeLogin(@RequestBody LoginUserInfo loginUserInfo) {
         throw new IllegalStateException("Add Spring Security to handle authentication");
-    };
+    }
 
     @ApiOperation("注销登录")
     @PostMapping("/logout")
     default void fakeLogout() {
         throw new IllegalStateException("Add Spring Security to handle authentication");
-    };
+    }
 
     @ApiOperation("当前登录用户")
-    Map<String, Object> current();
+    Map<String, Object> current(HttpServletRequest request);
 }
