+++ sb2-web-plat/src/main/java/com/incarcloud/sb2/auth/controller/AuthController.java
@@ -5,8 +5,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpServletRequest;
+import java.security.Principal;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 
 /**
  * 登录授权模块控制器
@@ -18,12 +21,23 @@ import java.util.Map;
 @RequestMapping("/api/plat/auth")
 public class AuthController implements AuthApi {
 
+    /**
+     * Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+     * if (!(authentication instanceof AnonymousAuthenticationToken)) {
+     *     String currentUserName = authentication.getName();
+     *     return currentUserName;
+     * }
+     */
+
     @Override
     @RequestMapping(value = "/current", method = RequestMethod.GET)
-    public Map<String, Object> current() {
+    public Map<String, Object> current(HttpServletRequest request) {
         Map<String, Object> returnData = new HashMap<>();
         returnData.put("code", "0000");
-        returnData.put("message", "authRedirect");
+        returnData.put("message", "current");
+
+        Optional<Principal> principal = Optional.ofNullable(request.getUserPrincipal());
+        returnData.put("data", principal.map(Principal::getName).orElse(null));
         return returnData;
     }
 }
