+++ sb2-web-plat/src/main/java/com/incarcloud/sb2/config/BizSwagger2Configuration.java
@@ -1,17 +1,11 @@
 package com.incarcloud.sb2.config;
 
-import com.incarcloud.sb2.security.Swagger2LoginOperations;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.spring.web.plugins.DocumentationPluginsManager;
-import springfox.documentation.spring.web.scanners.ApiDescriptionReader;
-import springfox.documentation.spring.web.scanners.ApiListingScanner;
-import springfox.documentation.spring.web.scanners.ApiModelReader;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 /**
@@ -22,7 +16,7 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
  */
 @Configuration
 @EnableSwagger2
-public class ProjectSwagger2Configuration extends DefaultSwagger2ConfigurationSupport {
+public class BizSwagger2Configuration extends DefaultSwagger2ConfigurationSupport {
 
     @Bean
     public Docket createRestApi() {
@@ -35,15 +29,4 @@ public class ProjectSwagger2Configuration extends DefaultSwagger2ConfigurationSu
                 .paths(PathSelectors.regex("/api/plat/.*"))
                 .build();
     }
-
-    /**
-     * Spring Security登录与注销操作API
-     *
-     * @return
-     */
-//    @Bean
-//    @Primary
-//    public ApiListingScanner addExtraOperations(ApiDescriptionReader apiDescriptionReader, ApiModelReader apiModelReader, DocumentationPluginsManager pluginsManager) {
-//        return new Swagger2LoginOperations(apiDescriptionReader, apiModelReader, pluginsManager);
-//    }
 }
