--- sb2-web-plat/src/main/java/com/incarcloud/sb2/security/UserDto.java
@@ -1,114 +1,0 @@
-package com.incarcloud.sb2.security;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import javax.management.relation.Role;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-/**
- * 登录用户详情
- *
- * @author Aaric, created on 2019-07-31T11:19.
- * @since 0.6.0-SNAPSHOT
- */
-public class UserDto implements UserDetails {
-
-    private Long id;
-    private String u;
-    private String p;
-    private Long ts;
-    private String v;
-
-    private List<Role> roleList;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        List<GrantedAuthority> authorityList = new ArrayList<>();
-        for (Role role : roleList) {
-            authorityList.add(new SimpleGrantedAuthority(role.getRoleName()));
-        }
-        return authorityList;
-    }
-
-    @Override
-    public String getPassword() {
-        return null;
-    }
-
-    @Override
-    public String getUsername() {
-        return null;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getU() {
-        return u;
-    }
-
-    public void setU(String u) {
-        this.u = u;
-    }
-
-    public String getP() {
-        return p;
-    }
-
-    public void setP(String p) {
-        this.p = p;
-    }
-
-    public Long getTs() {
-        return ts;
-    }
-
-    public void setTs(Long ts) {
-        this.ts = ts;
-    }
-
-    public String getV() {
-        return v;
-    }
-
-    public void setV(String v) {
-        this.v = v;
-    }
-
-    public List<Role> getRoleList() {
-        return roleList;
-    }
-
-    public void setRoleList(List<Role> roleList) {
-        this.roleList = roleList;
-    }
-}
